test: session share of contract fixtures
